+++ changes.md
@@ -1,1 +1,2 @@
-1. improve pagination, (move to start button) 
+1. clean up components, group the folders into tabs/features (e.g settings has voicepicer and modelpicker and settings)
+2. update architectural diagram